feat: streamline document service and controller
Adjusts the document service to return a future for document content like all other service methods.
Removes explicit exception handling from document controller and moves it into global exception handler.
Moves result mapping of document controller into response mapper.
