+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/ResponseMapper.java
@@ -17,6 +17,7 @@ import static java.time.temporal.ChronoField.SECOND_OF_MINUTE;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.document.api.DocumentLink;
+import io.camunda.service.DocumentServices.DocumentContentResponse;
 import io.camunda.service.DocumentServices.DocumentErrorResponse;
 import io.camunda.service.DocumentServices.DocumentReferenceResponse;
 import io.camunda.service.exception.ServiceException;
@@ -102,10 +103,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
+import org.springframework.http.InvalidMediaTypeException;
 import org.springframework.http.MediaType;
 import org.springframework.http.ProblemDetail;
 import org.springframework.http.ResponseEntity;
 import org.springframework.util.CollectionUtils;
+import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBody;
 
 public final class ResponseMapper {
 
@@ -258,6 +261,31 @@ public final class ResponseMapper {
             .tenantId(brokerResponse.getTenantId())
             .processInstanceKey(KeyUtil.keyToString(brokerResponse.getProcessInstanceKey()));
     return new ResponseEntity<>(response, HttpStatus.OK);
+  }
+
+  public static ResponseEntity<StreamingResponseBody> toDocumentContentResponse(
+      final DocumentContentResponse response) {
+    final MediaType mediaType = resolveMediaType(response);
+    return ResponseEntity.ok()
+        .contentType(mediaType)
+        .body(
+            bodyStream -> {
+              try (final var contentInputStream = response.content()) {
+                contentInputStream.transferTo(bodyStream);
+              }
+            });
+  }
+
+  private static MediaType resolveMediaType(final DocumentContentResponse contentResponse) {
+    try {
+      final var contentType = contentResponse.contentType();
+      if (contentType == null) {
+        return MediaType.APPLICATION_OCTET_STREAM;
+      }
+      return MediaType.parseMediaType(contentResponse.contentType());
+    } catch (final InvalidMediaTypeException e) {
+      return MediaType.APPLICATION_OCTET_STREAM;
+    }
   }
 
   public static ResponseEntity<Object> toDocumentReference(
