+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/controller/DocumentController.java
@@ -10,7 +10,6 @@ package io.camunda.zeebe.gateway.rest.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.security.auth.CamundaAuthenticationProvider;
 import io.camunda.service.DocumentServices;
-import io.camunda.service.DocumentServices.DocumentContentResponse;
 import io.camunda.service.DocumentServices.DocumentLinkParams;
 import io.camunda.zeebe.gateway.protocol.rest.DocumentLinkRequest;
 import io.camunda.zeebe.gateway.protocol.rest.DocumentMetadata;
@@ -23,8 +22,6 @@ import io.camunda.zeebe.gateway.rest.annotation.CamundaPostMapping;
 import jakarta.servlet.http.Part;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CompletionException;
-import org.springframework.http.InvalidMediaTypeException;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -96,51 +93,18 @@ public class DocumentController {
   @CamundaGetMapping(
       path = "/{documentId}",
       produces = {}) // produces arbitrary content type
-  public ResponseEntity<StreamingResponseBody> getDocumentContent(
+  public CompletableFuture<ResponseEntity<StreamingResponseBody>> getDocumentContent(
       @PathVariable final String documentId,
       @RequestParam(required = false) final String storeId,
       @RequestParam(required = false) final String contentHash) {
 
-    try {
-      final DocumentContentResponse contentResponse =
-          getDocumentContentResponse(documentId, storeId, contentHash);
-      final MediaType mediaType = resolveMediaType(contentResponse);
-      return ResponseEntity.ok()
-          .contentType(mediaType)
-          .body(
-              bodyStream -> {
-                try (final var contentInputStream = contentResponse.content()) {
-                  contentInputStream.transferTo(bodyStream);
-                }
-              });
-    } catch (final Exception e) {
-      // we can't return a generic Object type when streaming a response due to Spring MVC
-      // limitations
-      // exception handling is done in the exception handler below
-      if (e instanceof final CompletionException ce) {
-        throw new DocumentContentFetchException("Failed to get document content", ce.getCause());
-      }
-      throw new DocumentContentFetchException("Failed to get document content", e);
-    }
-  }
-
-  private MediaType resolveMediaType(final DocumentContentResponse contentResponse) {
-    try {
-      final var contentType = contentResponse.contentType();
-      if (contentType == null) {
-        return MediaType.APPLICATION_OCTET_STREAM;
-      }
-      return MediaType.parseMediaType(contentResponse.contentType());
-    } catch (final InvalidMediaTypeException e) {
-      return MediaType.APPLICATION_OCTET_STREAM;
-    }
-  }
-
-  private DocumentContentResponse getDocumentContentResponse(
-      final String documentId, final String storeId, final String contentHash) {
+    // handle the future explicitly here because a StreamingResponseBody is needed as result instead
+    // of Object
     return documentServices
         .withAuthentication(authenticationProvider.getCamundaAuthentication())
-        .getDocumentContent(documentId, storeId, contentHash);
+        .getDocumentContent(documentId, storeId, contentHash)
+        // Any service exception that can occur is handled by the GlobalControllerExceptionHandler
+        .thenApplyAsync(ResponseMapper::toDocumentContentResponse);
   }
 
   @CamundaDeleteMapping(path = "/{documentId}")
@@ -179,12 +143,5 @@ public class DocumentController {
                 .withAuthentication(authenticationProvider.getCamundaAuthentication())
                 .createLink(documentId, storeId, contentHash, params),
         ResponseMapper::toDocumentLinkResponse);
-  }
-
-  public static class DocumentContentFetchException extends RuntimeException {
-
-    public DocumentContentFetchException(final String message, final Throwable cause) {
-      super(message, cause);
-    }
   }
 }
